+++ game/src/pages/login/[[...index]].tsx
@@ -2,7 +2,7 @@ import { SignIn, SignedOut } from "@clerk/nextjs";
 
 import { type NextPage } from "next";
 
-import Layout from "@/components/Layout/Layout";
+import Layout from "@/components/common/Layout/Layout";
 
 const Login: NextPage = () => {
   return (
