+++ game/src/pages/military.tsx
@@ -2,8 +2,8 @@ import { useUser } from "@clerk/nextjs";
 
 import { type NextPage } from "next";
 
-import Layout from "@/components/Layout/Layout";
-import Military from "@/components/Military/Military";
+import Layout from "@/components/common/Layout/Layout";
+import Military from "@/components/features/Military/Military";
 
 import { api } from "@/utils/api";
