+++ game/src/pages/resources.tsx
@@ -2,13 +2,15 @@ import { useUser } from "@clerk/nextjs";
 
 import { type NextPage } from "next";
 
-import Layout from "@/components/Layout/Layout";
-import ResourceTable from "@/components/Resources/ResourceTable";
-import LoadingSpinner from "@/components/Loader/LoadingSpinner";
-import BarGraph from "@/components/Resources/BarGraph";
+import Layout from "@/components/common/Layout/Layout";
+import LoadingSpinner from "@/components/common/Loader/LoadingSpinner";
+import BarGraph from "@/components/features/Resources/BarGraph";
+import ResourceTable from "@/components/features/Resources/ResourceTable";
+
 
 import { api } from "@/utils/api";
 import { renderIncomeData } from "@/utils/functions";
+
 
 const Resources: NextPage = () => {
   const { user, isSignedIn, isLoaded } = useUser();
